green

+++ src/main/java/de/jlab/scales/anki/Card.java
@@ -1,12 +1,14 @@
 package de.jlab.scales.anki;
 
+import de.jlab.scales.difficulty.WithDifficulty;
+
 public interface Card extends WithDifficulty, WithAssets, Comparable<Card> {
 
   String getCsv();
   String getHtml();
   
   default int compareTo(Card that) {
-    return Integer.compare(this.getDifficulty(), that.getDifficulty());
+    return Double.compare(this.getDifficulty(), that.getDifficulty());
   }
 
 }
